Membuat authorization, menambahkan package spatie laravel log
* Menambahkan tabel
  - activity_log
  - permission (hak akses)
  - level_permission (hak akses role/level)
* Memodifikasi navbar (pelanggan, dan admin)
* Membuat authorization
  - Membuat Middleware AuthGates
  - Membuat Middleware IsAdmin
  - Membuat Gates
* Menambahkan pengecekan autentikasi user di method index LoginController
  Ilustrasi: (user->isLogin ? redirect()->intended() : login)
* Memodifikasi halaman riwayat pembayaran
* Membuat folder "pelanggan" untuk menyimpan halaman-halaman pelanggan
* Menambahkan soft deletes di semua tabel master
* Menghapus plugin responsive + button datatables
* Membenarkan beberapa logika dan algoritma yang salah
* Membenarkan relasi model
* Membuat model PaymentHistory
* Memodifikasi seeder dan menambahkan seeder baru (PermissionSeeder dan PermissionLevelSeeder)
* Mengubah rute path "Home", '/home' -> '/'
* Menambahkan form request PlnCustomerRequest, UserRequest, dan UsageRequest
* Membuat auto login setelah registrasi

+++ app/Http/Controllers/Admin/ElectricityUsageController.php
@@ -3,12 +3,19 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\UsageRequest;
 use App\Models\Bill;
 use App\Models\PlnCustomer;
 use App\Models\Usage;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Yajra\DataTables\Facades\DataTables;
+use Symfony\Component\HttpFoundation\Response;
+
+/**
+ * Resource Controller untuk model Usage
+ */
 
 class ElectricityUsageController extends Controller
 {
@@ -19,14 +26,17 @@ class ElectricityUsageController extends Controller
      */
     public function index(Request $request)
     {
+        // dd(activity()->log('Look mum, I logged something'));
+        abort_if(Gate::denies('usage_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         if($request->ajax()){
             $usage = Usage::all();
             return DataTables::of($usage)
                     ->addColumn('action', function($usage){
-                        $button = '<a href='. route("admin.usage.edit", $usage->id).' class="btn btn-success btn-sm">edit</a>';
-                        $button .= '<a href='. route("admin.usage.show", $usage->id).' class="btn btn-primary btn-sm mx-2">detail</a>';
+                        $button = '<a href='. route("admin.usages.edit", $usage->id).' class="btn btn-success btn-sm">edit</a>';
+                        $button .= '<a href='. route("admin.usages.show", $usage->id).' class="btn btn-primary btn-sm mx-2">detail</a>';
                         $button .= '
-                            <form action='.route("admin.usage.destroy", $usage->id).' method="POST" class="d-inline-block form-delete">
+                            <form action='.route("admin.usages.destroy", $usage->id).' method="POST" class="d-inline-block form-delete">
                                 '. csrf_field() .'
                                 '. method_field("DELETE") .'
                                 <button type="submit" class="btn btn-danger btn-sm btn-delete">delete</button>
@@ -46,6 +56,7 @@ class ElectricityUsageController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('usage_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $customers = PlnCustomer::get();
         return view('pages.admin.electricity-usage.create', compact('customers'));
     }
@@ -53,13 +64,13 @@ class ElectricityUsageController extends Controller
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Http\Requests\UsageRequest  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(UsageRequest $request)
     {
-        Usage::create($request->except('_token'));
-        return redirect()->route('admin.usage.index')->withSuccess('Data berhasil ditambahkan!');
+        Usage::create($request->all());
+        return redirect()->route('admin.usages.index')->withSuccess('Data berhasil ditambahkan!');
     }
 
     /**
@@ -70,6 +81,7 @@ class ElectricityUsageController extends Controller
      */
     public function show(Usage $usage)
     {
+        abort_if(Gate::denies('usage_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return view('pages.admin.electricity-usage.show', compact('usage'));
     }
 
@@ -81,6 +93,7 @@ class ElectricityUsageController extends Controller
      */
     public function edit(Usage $usage)
     {
+        abort_if(Gate::denies('usage_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $customers = PlnCustomer::get();
         return view('pages.admin.electricity-usage.edit', compact('usage', 'customers'));
     }
@@ -88,14 +101,15 @@ class ElectricityUsageController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Http\Requests\UsageRequest  $request
      * @param  \App\Models\Usage  $usage
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Usage $usage)
+    public function update(UsageRequest $request, Usage $usage)
     {
-        $usage->update($request->except(['_token', 'method']));
-        return redirect()->route('admin.usage.index')->withSuccess('Data penggunaan berhasil diubah!');
+        abort_if(Gate::denies('usage_update'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        $usage->update($request->all());
+        return redirect()->route('admin.usages.index')->withSuccess('Data penggunaan berhasil diubah!');
     }
 
     /**
@@ -106,10 +120,23 @@ class ElectricityUsageController extends Controller
      */
     public function destroy(Usage $usage)
     {
-        $bill = Bill::where('id_penggunaan', $usage->id)->first();
-        $bill->usage()->dissociate();
-        $bill->save();
+        abort_if(Gate::denies('usage_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        /**
+         * jika penggunaan ini memiliki relasi dengan tagihan, 
+         * dan penggunaan ini memiliki relasi dengan pelanggan PLN, 
+         * maka soft deletes.
+         */
+        if($usage->bill_count > 0 && $usage->pln_customer_count > 0){
+            $bill = Bill::firstWhere('id_penggunaan', $usage->id);
+            $bill->usage()->dissociate();
+            $bill->save();
+
+            $usage->plnCustomer()->dissociate();
+            $usage->save();
+        }
+
         $usage->delete();
-        return redirect()->route('admin.usage.index')->withSuccess('Data penggunaan berhasil dihapus!');
+        return redirect()->route('admin.usages.index')->withSuccess('Data penggunaan berhasil dihapus!');
     }
 }
